feat(web-domain-reseller): datagrid
ref: #DCE-163
