+++ packages/manager/apps/web-domains/src/common/constants.ts
@@ -27,3 +27,8 @@ export const RENEW_URL: Record<string, string> = {
 };
 
 export const DOMAIN = 'DOMAIN';
+
+export const domainsListApiUrl = '/domain/name';
+export const resellerDomainsListApiUrl = (contactBilling: string) => {
+  return `/domain/name?contactBilling=${contactBilling}`;
+};
