+++ packages/manager/apps/web-domains/src/domain-reseller/pages/tabs/DomainsList.tsx
@@ -1,3 +1,31 @@
+import DomainsList from '@/common/components/DomainsList/domainsList';
+import { resellerDomainsListApiUrl } from '@/common/constants';
+import {
+  ServiceInfoContactEnum,
+  ServiceRoutes,
+} from '@/common/enum/common.enum';
+import { useGetServiceInformationByRoutes } from '@/common/hooks/data/query';
+import { findContact } from '@/common/utils/utils';
+import Loading from '@/domain/components/Loading/Loading';
+import { useMemo } from 'react';
+
 export default function DomainResellerDomainsList() {
-  return 'TODO DCE-163';
+  const {
+    serviceInfo,
+    isServiceInfoLoading,
+  } = useGetServiceInformationByRoutes(ServiceRoutes.DomainReseller);
+
+  const nicAdminReseller = useMemo(() => {
+    if (!serviceInfo?.customer?.contacts) return undefined;
+    return findContact(
+      serviceInfo.customer.contacts,
+      ServiceInfoContactEnum.Administrator,
+    );
+  }, [serviceInfo]);
+
+  if (isServiceInfoLoading) {
+    return <Loading />;
+  }
+
+  return <DomainsList baseRoute={resellerDomainsListApiUrl(nicAdminReseller)} />;
 }
